refactor: Dissolved 'services/events' folder into 'services' folder in core module.
