+++ src/app/core/core.module.ts
@@ -6,7 +6,7 @@ import { HttpClientModule } from '@angular/common/http'
 import { HttpBasicRundownService } from './services/http-basic-rundown.service'
 import { HttpRundownService } from './services/http-rundown.service'
 import { HttpAdLibPieceService } from './services/http-ad-lib-piece.service'
-import { RundownEventObserver } from './services/events/rundown-event-observer.service'
+import { RundownEventObserver } from './services/rundown-event-observer.service'
 import { RundownStateService } from './services/rundown-state.service'
 import { RundownEventParser } from './interfaces/rundown-event-parser.interface'
 import { BasicRundownStateService } from './services/basic-rundown-state.service'
